Split input system
