+++ crates/lavagna-bevy/src/lib.rs
@@ -1,9 +1,12 @@
+mod input;
+
 use bevy::{
     diagnostic::{Diagnostics, FrameTimeDiagnosticsPlugin},
-    input::{mouse::*, ButtonState},
     prelude::*,
     window::Window,
 };
+
+use crate::input::*;
 
 pub fn run() {
     App::new()
@@ -15,8 +18,7 @@ pub fn run() {
             ..default()
         }))
         .add_startup_system(setup)
-        .add_system(mouse_system)
-        .add_system(cursor_system)
+        .add_system(input_system)
         .add_system(debug_test_system)
         .run();
 }
@@ -49,50 +51,6 @@ fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
         }),
         DebugText,
     ));
-}
-
-fn mouse_system(
-    mut mouse_button_input_events: EventReader<MouseButtonInput>,
-    mut pen_q: Query<&mut Pen, With<LocalPen>>,
-) {
-    let mut pen = pen_q.single_mut();
-
-    for event in mouse_button_input_events.iter() {
-        match event {
-            MouseButtonInput {
-                button: MouseButton::Left,
-                state: ButtonState::Pressed,
-            } => {
-                pen.pressed = true;
-            }
-            MouseButtonInput {
-                button: MouseButton::Left,
-                state: ButtonState::Released,
-            } => {
-                pen.pressed = false;
-            }
-            _ => {}
-        }
-    }
-}
-
-fn cursor_system(
-    window_q: Query<&Window>,
-    camera_q: Query<(&Camera, &GlobalTransform), With<MainCamera>>,
-    mut pen_q: Query<&mut Pen, With<LocalPen>>,
-) {
-    let (camera, camera_transform) = camera_q.single();
-    let mut pen = pen_q.single_mut();
-    let window = window_q.single();
-
-    if let Some(world_position) = window
-        .cursor_position()
-        .and_then(|cursor| camera.viewport_to_world(camera_transform, cursor))
-        .map(|ray| ray.origin.truncate())
-    {
-        pen.x = world_position[0] as i64;
-        pen.y = world_position[1] as i64;
-    }
 }
 
 #[derive(Component)]
